Boolean searches
Add method for running a search. This reuses the TermInfo struct for storing
intermediate results when processing the query, so combinations of terms the
same way as single terms. A subquery like "A and B" is as if we just made up a
word for "A and B" and counted occurances of that in the index.