Add support for OR queries
Writing `foo AND bar OR baz` means `foo AND (bar or baz)` not
`(foo AND bar) OR baz`.

I think this is the more natural interpretation (searching for variations on a
word rather than variations on a phrase) and I don't want to parse
arbitrary boolean queries with brackets - just AND/OR/NOT will do.